% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addClonalSequence}
\alias{addClonalSequence}
\title{Adds columns to DB, containing the consensus clonal sequence}
\usage{
addClonalSequence(db, sequenceColumn = "SEQUENCE_GAP",
  germlineColumn = "GERMLINE_GAP_D_MASK")
}
\arguments{
\item{db}{a data.frame of the DB file.}

\item{sequenceColumn}{The name of the sequence column.}

\item{germlineColumn}{The name of the germline column.}
}
\value{
db  a data.frame of the DB file
}
\description{
This identifies the consensus clonal sequence to the DB file.
}

