% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addObservedMutations}
\alias{addObservedMutations}
\title{Adds columns to DB, containing the numbers of observed mutations}
\usage{
addObservedMutations(db, sequenceColumn = "SEQUENCE_GAP",
  germlineColumn = "GERMLINE_GAP_D_MASK")
}
\arguments{
\item{db}{a data.frame of the DB file.}

\item{sequenceColumn}{The name of the sequence column.}

\item{germlineColumn}{The name of the germline column.}
}
\value{
db  a data.frame of the DB file
}
\description{
This adds the observed number of mutations to the DB file.\cr
}

