% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcGroupedBaseline}
\alias{calcGroupedBaseline}
\title{Calculated CDR and FWR BASELINe selection strengths}
\usage{
calcGroupedBaseline(db, columnsToGroupBy)
}
\arguments{
\item{db}{a data.frame of the DB file.}
}
\value{
a data.frame of the BASELINe sigma values and 95% Conf. Intervals for the CDR and FWR
}
\description{
Given a DB file contianing the columns "CDR_PDFs" and "FWR_PDFs".\cr
this function groups the sequences by specified columns
}

