% Generated by roxygen2 (4.0.2): do not edit by hand
\name{computeBaselinePDF}
\alias{computeBaselinePDF}
\title{Adds columns to DB, containing the BASELINE prob. density functions}
\usage{
computeBaselinePDF(db)
}
\arguments{
\item{db}{a data.frame of the DB file.}
}
\value{
db  a data.frame of the DB file
}
\description{
This adds the BASELINE prob. density functions to the DB file.\cr
}

