% Generated by roxygen2 (4.0.2): do not edit by hand
\name{computeMutationTypes}
\alias{computeMutationTypes}
\title{Adds columns to DB, containing the expected frequencies of mutations}
\usage{
computeMutationTypes(param_strSeq)
}
\arguments{
\item{param_strSeq}{The name of the sequence column.}

\item{param_vecMutabilities}{The name of the germline column.}
}
\value{
matTargeting
}
\description{
This adds the expected frequencies number of mutations to the DB file.\cr
}

