% Generated by roxygen2 (4.0.2): do not edit by hand
\name{countMutations}
\alias{countMutations}
\title{Count number of mutations in sequence}
\usage{
countMutations(seqInput, seqGL)
}
\arguments{
\item{seqInput}{Input sequence}

\item{seqGL}{Germline sequence}
}
\value{
array of observed mutations
}
\description{
This function counts the number of mutations in a sequence.\cr
}

