% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createMutabilityModel}
\alias{createMutabilityModel}
\title{Builds a mutability model}
\usage{
createMutabilityModel(db, model = "RS",
  substitutionModel = substitutionModel, sequenceColumn = "SEQUENCE_GAP",
  germlineColumn = "GERMLINE_GAP_D_MASK", vCallColumn = "V_CALL_GENOTYPED",
  multipleMutation = 0)
}
\arguments{
\item{db}{a data.frame of the DB file.}

\item{model}{The type of model to create "S" = silent, "RS" = Silent = Replacement mutations}

\item{substitutionModel}{matrix of fivemers and the substitution counts (from createSubstitutionModel)}

\item{sequenceColumn}{The name of the sequence column}

\item{germlineColumn}{The name of the germline column}

\item{multipleMutation}{0 = Treat independently, 1 = Ignore codons with multiple mutations}

\item{VCallColumn}{The name of the genotyped columnn}
}
\value{
list of fivemers and the substitution counts
}
\description{
This builds a 5-mer based mutability model.\cr
}

