% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionDefinitions.R
\name{createRegionDefinition}
\alias{createRegionDefinition}
\title{Creates a RegionDefinition}
\usage{
createRegionDefinition(name = "", boundaries = factor(), description = "",
  citation = "")
}
\arguments{
\item{name}{name of the region definition.}

\item{boundaries}{\code{factor} defining the region boundaries of the sequence.
The levels and values of \code{boundaries} determine the 
number of regions (e.g. CDR and FWR).}

\item{description}{description of the region definition and its source data.}

\item{citation}{publication source.}
}
\value{
A \code{RegionDefinition} object.
}
\description{
\code{createRegionDefinition} creates a \code{RegionDefinition}.
}
\examples{
# Creates an empty RegionDefinition object
createRegionDefinition()

}
\seealso{
See \link{RegionDefinition} for the return object.
}
