% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createSubstitutionModel}
\alias{createSubstitutionModel}
\title{Builds a substitution model}
\usage{
createSubstitutionModel(db, model = "RS", sequenceColumn = "SEQUENCE_GAP",
  germlineColumn = "GERMLINE_GAP_D_MASK", vCallColumn = "V_CALL_GENOTYPED",
  multipleMutation = 0)
}
\arguments{
\item{db}{a data.frame of the DB file.}

\item{model}{The type of model to create "S" = silent, "RS" = Silent = Replacement mutations}

\item{sequenceColumn}{The name of the sequence column}

\item{germlineColumn}{The name of the germline column}

\item{multipleMutation}{0 = Treat independently, 1 = Ignore codons with multiple mutations}

\item{VCallColumn}{The name of the genotyped columnn}
}
\value{
list of fivemers and the substitution counts
}
\description{
This builds a 5-mer based substitution model.\cr
}

