% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistToNearest.R
\docType{class}
\name{DensityThreshold-class}
\alias{DensityThreshold-class}
\alias{DensityThreshold}
\alias{print,DensityThreshold-method}
\alias{DensityThreshold-method}
\alias{plot,DensityThreshold,missing-method}
\alias{DensityThreshold-method}
\title{Output of the \code{dens} method of findThreshold}
\usage{
\S4method{print}{DensityThreshold}(x)

\S4method{plot}{DensityThreshold,missing}(x, y, ...)
}
\arguments{
\item{x}{DensityThreshold object}

\item{y}{ignored.}

\item{...}{arguments to pass to \link{plotDensityThreshold}.}
}
\description{
\code{DensityThreshold} contains output from the \code{dens} method \link{findThreshold}.
}
\section{Slots}{

\describe{
\item{\code{x}}{input distance vector with NA or infinite values removed.}

\item{\code{bandwidth}}{bandwidth value fit during density estimation.}

\item{\code{xdens}}{x-axis (distance value) vector for smoothed density estimate.}

\item{\code{ydens}}{y-axis (density) vector for smoothed density estimate.}

\item{\code{threshold}}{distance threshold that separates two modes of the input distribution.}
}}

\seealso{
\link{findThreshold}
}
