% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{getnproc}
\alias{getnproc}
\title{Determines the numbers of CPU cores available}
\usage{
getnproc()
}
\value{
The number of cores available. Returns 1 if undeterminable.
}
\description{
Determines the numbers of CPU cores available
}
\examples{
getnproc()

}
