% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{getPlatform}
\alias{getPlatform}
\title{Determines the OS platform being used}
\usage{
getPlatform()
}
\value{
The OS platform.
}
\description{
Determines the OS platform being used
}
\examples{
getPlatform()

}
