% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetingModels.R
\docType{data}
\name{HH_S1F}
\alias{HH_S1F}
\title{Human heavy chain, silent, 1-mer, functional substitution model.}
\format{A 4x4 matrix of nucleotide substitution rates. The rates are normalized,
          therefore each row sums up to 1.}
\usage{
HH_S1F
}
\description{
1-mer substitution model of somatic hypermutation based on analysis of silent mutations
in functional heavy chain Ig sequences from Homo sapiens.
}
\note{
\code{HH_S1F} replaces \code{HS1FDistance} in versions of SHazaM prior to 0.1.5.
}
\references{
\enumerate{
  \item  Yaari G, et al. Models of somatic hypermutation targeting and substitution based 
           on synonymous mutations from high-throughput immunoglobulin sequencing data. 
           Front Immunol. 2013 4(November):358.
}
}
\seealso{
See \link{HKL_S1F} for the human light chain 1-mer substitution model and 
          \link{MK_RS1NF} for the mouse light chain 1-mer substitution model.
}
\keyword{datasets}
