% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetingModels.R
\docType{data}
\name{HH_S5F}
\alias{HH_S5F}
\title{Human heavy chain, silent, 5-mer, functional targeting model.}
\format{A \link{TargetingModel} object.}
\usage{
HH_S5F
}
\description{
5-mer model of somatic hypermutation targeting based on analysis of silent mutations
in functional heavy chain Ig sequences from Homo sapiens.
}
\references{
\enumerate{
  \item  Yaari G, et al. Models of somatic hypermutation targeting and substitution based 
           on synonymous mutations from high-throughput immunoglobulin sequencing data. 
           Front Immunol. 2013 4(November):358.
 }
}
\seealso{
See \link{HH_S1F} for the 1-mer substitution matrix from the same 
          publication; \link{HKL_S5F} for the human light chain 5-mer targeting model; 
          \link{MK_RS5NF} for the mouse 5-mer targeting model; and \link{U5N} for the 
          uniform 5-mer null targeting model.
}
\keyword{datasets}
