% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetingModels.R
\docType{data}
\name{HKL_S1F}
\alias{HKL_S1F}
\title{Human kappa and lambda chain, silent, 1-mer, functional substitution model.}
\format{A 4x4 matrix of nucleotide substitution rates. The rates are normalized,
          therefore each row sums up to 1.}
\usage{
HKL_S1F
}
\description{
1-mer substitution model of somatic hypermutation based on analysis of silent mutations
in functional kappa and lambda light chain Ig sequences from Homo sapiens.
}
\note{
Reported in Table III in Cui et al, 2016.
}
\references{
\enumerate{
  \item  Cui A, Di Niro R, Vander Heiden J, Briggs A, Adams K, Gilbert T, O'Connor K,
  Vigneault F, Shlomchik M and Kleinstein S (2016). A Model of Somatic Hypermutation 
  Targeting in Mice Based on High-Throughput Ig Sequencing Data. The Journal of 
  Immunology, 197(9), 3566-3574.
 }
}
\seealso{
See \link{HH_S1F} for the human heavy chain 1-mer substitution model and 
          \link{MK_RS1NF} for the mouse light chain 1-mer substitution model.
}
\keyword{datasets}
