% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetingModels.R
\docType{data}
\name{HKL_S5F}
\alias{HKL_S5F}
\title{Human kappa and lambda light chain, silent, 5-mer, functional targeting model.}
\format{A \link{TargetingModel} object.}
\usage{
HKL_S5F
}
\description{
5-mer model of somatic hypermutation targeting based on analysis of silent mutations
in functional kappa and lambda light chain Ig sequences from Homo sapiens.
}
\references{
\enumerate{
  \item  Cui A, Di Niro R, Vander Heiden J, Briggs A, Adams K, Gilbert T, O'Connor K,
  Vigneault F, Shlomchik M and Kleinstein S (2016). A Model of Somatic Hypermutation 
  Targeting in Mice Based on High-Throughput Ig Sequencing Data. The Journal of 
  Immunology, 197(9), 3566-3574.
 }
}
\seealso{
See \link{HH_S5F} for the human heavy chain 5-mer targeting model; 
          \link{MK_RS5NF} for the mouse kappa light chain 5-mer targeting model; 
          and \link{U5N} for the uniform 5-mer null targeting model.
}
\keyword{datasets}
