% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetingModels.R
\docType{data}
\name{HS1FDistance}
\alias{HS1FDistance}
\title{Human single nucleotide distance matrix.}
\format{A symmetric matrix of nucleotide substitution distance scores with 
          row names and column names definition the specific subsitution.}
\usage{
HS1FDistance
}
\description{
Single nucleotide distance matrix of somatic hypermutation targeting based on 
human Ig sequence data.
}
\references{
\enumerate{
  \item  Yaari G, et al. Models of somatic hypermutation targeting and substitution based 
           on synonymous mutations from high-throughput immunoglobulin sequencing data. 
           Front Immunol. 2013 4(November):358.
}
}
\seealso{
See \code{\link{HS5FModel}} for the 5-mer model from same publication.
}
\keyword{datasets}

