% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\docType{data}
\name{InfluenzaDb}
\alias{InfluenzaDb}
\title{Example Change-O database}
\format{A data.frame containing Change-O style data.}
\usage{
InfluenzaDb
}
\description{
A small example database subset from Laserson and Vigneault et al, 2014.
}
\references{
\enumerate{
  \item  Laserson U and Vigneault F, et al. High-resolution antibody dynamics of 
           vaccine-induced immune responses. 
           Proc Natl Acad Sci USA. 2014 111:4928-33.
}
}
\seealso{
See \code{\link{HS1FDistance}} for the human 1-mer distance matrix.
}
\keyword{datasets}

