% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listObservedMutations}
\alias{listObservedMutations}
\title{List the numbers of observed mutations}
\usage{
listObservedMutations(db, sequenceColumn = "SEQUENCE_GAP",
  germlineColumn = "GERMLINE_GAP_D_MASK")
}
\arguments{
\item{db}{a data.frame of the DB file.}

\item{sequenceColumn}{The name of the sequence column.}

\item{germlineColumn}{The name of the germline column.}
}
\value{
list of mutations in each clone
}
\description{
This lists the observed number of mutation.\cr
}

