% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadDB}
\alias{loadDB}
\title{Load DB file}
\usage{
loadDB(input_file)
}
\arguments{
\item{input_file}{Full path to tab-delimitted input file}
}
\value{
a data.frame of the DB file
}
\description{
This function loads the basic (CLIP) DB file.\cr
This file should be tab delimited and contain at least the following three columns: \cr
1. SEQUENCE_ID \cr
2. SEQUENCE_GAP \cr
3. GERMLINE_GAP (or GERMLINE_GAP_D_MASK) \cr
}
\details{
Note that the sequences need to gapped according to the IMGT numbering scheme.
}

