% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetingModels.R
\docType{data}
\name{M1NDistance}
\alias{M1NDistance}
\title{Mouse single nucleotide distance matrix.}
\format{A symmetric matrix of nucleotide substitution distance scores with 
          row names and column names definition the specific subsitution.}
\usage{
M1NDistance
}
\description{
Single nucleotide distance matrix of somatic hypermutation targeting based on 
Mus musculus Ig sequence data.
}
\references{
\enumerate{
  \item  Smith DS, et al. Di- and trinucleotide target preferences of somatic 
           mutagenesis in normal and autoreactive B cells. 
           J Immunol. 1996 156:2642-52. 
}
}
\seealso{
See \code{\link{HS1FDistance}} for the human 1-mer distance matrix.
}
\keyword{datasets}

