% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetingModels.R
\docType{data}
\name{MK_RS1NF}
\alias{MK_RS1NF}
\title{Mouse kappa chain, replacement and silent, 1-mer, non-functional substitution model.}
\format{A 4x4 matrix of nucleotide substitution rates. The rates are normalized,
          therefore each row sums up to 1.}
\usage{
MK_RS1NF
}
\description{
1-mer substitution model of somatic hypermutation based on analysis of replacement and
silent mutations in non-functional kappa light chain Ig sequences from NP-immunized Mus
musculus.
}
\note{
\code{MK_RS1NF} replaces \code{M1NDistance} from versions of SHazaM prior to 0.1.5.
}
\references{
\enumerate{
  \item  Cui A, Di Niro R, Vander Heiden J, Briggs A, Adams K, Gilbert T, O'Connor K,
  Vigneault F, Shlomchik M and Kleinstein S (2016). A Model of Somatic Hypermutation 
  Targeting in Mice Based on High-Throughput Ig Sequencing Data. The Journal of 
  Immunology, 197(9), 3566-3574.
 }
}
\seealso{
See \link{HH_S1F} for the human heavy chain 1-mer substitution model and 
          \link{HKL_S1F} for the human light chain 1-mer substitution model.
}
\keyword{datasets}
