% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetingModels.R
\docType{data}
\name{MK_RS5NF}
\alias{MK_RS5NF}
\title{Mouse kappa light chain, replacement and silent, 5-mer, non-functional targeting model.}
\format{\link{TargetingModel} object.}
\usage{
MK_RS5NF
}
\description{
5-mer model of somatic hypermutation targeting based on analysis of replacement and
silent mutations in non-functional kappa light chain Ig sequences from NP-immunized 
Mus musculus.
}
\references{
\enumerate{
  \item  Cui A, Di Niro R, Vander Heiden J, Briggs A, Adams K, Gilbert T, O'Connor K,
  Vigneault F, Shlomchik M and Kleinstein S (2016). A Model of Somatic Hypermutation 
  Targeting in Mice Based on High-Throughput Ig Sequencing Data. The Journal of 
  Immunology, 197(9), 3566-3574.
 }
}
\seealso{
See \link{MK_RS1NF} for the 1-mer substitution matrix from the same
          publication; \link{HH_S5F} for the human heavy chain silent 5-mer 
          functional targeting model; \link{HKL_S5F} for the human light chain 
          silent 5-mer functional targeting model; and \link{U5N} for the 
          uniform 5-mer null targeting model.
}
\keyword{datasets}
