% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetingModels.R
\docType{data}
\name{MRS5NFModel}
\alias{MRS5NFModel}
\title{Mouse 5-mer targeting model.}
\format{\link{TargetingModel} object.}
\usage{
MRS5NFModel
}
\description{
5-mer model of somatic hypermutation targeting based on analysis of replacement and
silent mutations in non-functional IgK from NP-immunized Mus musculus.
}
\references{
\enumerate{
  \item  Cui A, et al. A model of somatic hypermutation targeting in mice based on 
         high-throughput immunoglobulin sequencing data. Under review.
 }
}
\seealso{
See \link{HS5FModel} for human 5-mer targeting model, \link{U5NModel}
          for the uniform 5-mer model, and \link{M1NDistance} for the mouse 
          1-mer distance matrix.
}
\keyword{datasets}

