% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{cols}{number of columns in layout}

\item{layout}{a matrix specifying the layout. If present, 'cols' is ignored.}

\item{ggplot}{ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)}
}
\description{
Multiple plot function
}
\references{
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)
}

