% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationDefinitions.R
\name{MUTATION_SCHEMES}
\alias{CHARGE_MUTATIONS}
\alias{HYDROPATHY_MUTATIONS}
\alias{MUTATION_SCHEMES}
\alias{POLARITY_MUTATIONS}
\title{Amino acid mutation definitions}
\format{A \code{\link{MutationDefinition}} object defining:
\itemize{
  \item  \code{CHARGE_MUTATIONS}:      Amino acid mutations are defined by changes
                                       in charge classes.
  \item  \code{HYDROPATHY_MUTATIONS}:  Amino acid mutations are defined by changes
                                       in hydrophobicitity classes.
  \item  \code{POLARITY_MUTATIONS}:    Amino acid mutations are defined by changes
                                       in polarity classes.
}}
\description{
Definitions of replacement (R) and silent (S) mutations for different amino acid
physicochemical classes.
}
\references{
\enumerate{
  \item  \url{http://www.imgt.org/IMGTeducation/Aide-memoire/_UK/aminoacids/IMGTclasses.html} 
}
}

