% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotHedgehog}
\alias{plotHedgehog}
\title{Plot Mutabilities in Hedgehog circle format}
\usage{
plotHedgehog(mutabilityModel, outputName = "sample",
  outputPath = "/home/mu37")
}
\arguments{
\item{mutabilityModel}{matrix of mutabilities from createMutabilityModel()}

\item{outputName}{Filename for the output files}

\item{outputPath}{Path to save output}
}
\description{
Plot Mutabilities in Hedgehog circle format
}

