% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotSelection}
\alias{plotSelection}
\title{Plot the BASELINe selection values}
\usage{
plotSelection(db, groups = NULL, colors = NULL, outputName = NULL,
  outputPath = NULL)
}
\arguments{
\item{db}{data.frame returned by calcGroupedBaseline}

\item{groups}{column names by which you want the anlaysis grouped by}

\item{colors}{values are colors to assign to those group values}

\item{outputName}{Filename for the output files}

\item{outputPath}{Path to save output}
}
\description{
Plot the BASELINe selection values
}

