% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prepareSequence}
\alias{prepareSequence}
\title{Prepare SEQUENCE_GAP column}
\usage{
prepareSequence(db, columnName = c("SEQUENCE_GAP", "GERMLINE_GAP_D_MASK"))
}
\arguments{
\item{db}{a data.frame of the DB file}

\item{columnName}{The name of the column to prepare.}
}
\value{
a data.frame of the DB file
}
\description{
This function prepares the SEQUENCE_GAP column of the DB object.\cr
If the DB object uses a different name for the column or you need to prepare a different column please specify the column to  prepare.\cr
}

