% Generated by roxygen2 (4.0.2): do not edit by hand
\name{processNucMutations}
\alias{processNucMutations}
\title{Count mutations by region}
\usage{
processNucMutations(mutations)
}
\arguments{
\item{mutations}{an array of mutations. The name indicates the position and the values are the kinds of mutations.}
}
\value{
retArrary  an array of R_CDR, S_CDR, R_FWR and S_FWR
}
\description{
This function counts the number of mutations in a sequence and groups them by region.\cr
}

