% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shazam.R
\docType{package}
\name{shazam-package}
\alias{shazam-package}
\title{shazam: Immunoglobulin Somatic Hypermutation Analysis}
\description{
Provides a computational framework for analyzing mutations in immunoglobulin (Ig) sequences. Includes methods for Bayesian estimation of antigen-driven selection pressure, mutational load quantification, building of somatic hypermutation (SHM) models, and model-dependent distance calculations. Also includes empirically derived models of SHM for both mice and humans. Citations: Gupta and Vander Heiden, et al (2015) \doi{10.1093/bioinformatics/btv359}, Yaari, et al (2012) \doi{10.1093/nar/gks457}, Yaari, et al (2013) \doi{10.3389/fimmu.2013.00358}, Cui, et al (2016) \doi{10.4049/jimmunol.1502263}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shazam.readthedocs.io}
  \item Report bugs at \url{https://github.com/immcantation/shazam/issues}
}

}
\author{
\strong{Maintainer}: Susanna Marquez \email{susanna.marquez@yale.edu}

Authors:
\itemize{
  \item Susanna Marquez \email{susanna.marquez@yale.edu}
  \item Mohamed Uduman \email{mohamed.uduman@yale.edu}
  \item Namita Gupta \email{namita.gupta@yale.edu}
  \item Julian Zhou \email{julian.zhou@yale.edu}
  \item Nima Nouri \email{nima.nouri@yale.edu}
  \item Noah Yann Lee \email{noah.yann.lee@yale.edu}
  \item Jason Vander Heiden \email{jason.vanderheiden@gmail.com}
  \item Gur Yaari \email{gur.yaari@biu.ac.il}
  \item Steven Kleinstein \email{steven.kleinstein@yale.edu} [copyright holder]
}

Other contributors:
\itemize{
  \item Ang Cui \email{angcui@mit.edu} [contributor]
}

}
\keyword{internal}
