% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{shm}
\alias{shm}
\alias{shm-package}
\title{SHM}
\description{
Provides tools for advanced anaylisis of Ig Somatic HyperMuation. Includes BASELINe,
a novel method for quantifying selection in high-throughput Immunoglobulin sequencing
 data sets.
}
\references{
BASELINe
}

