% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetingModels.R
\docType{class}
\name{TargetingMatrix-class}
\alias{TargetingMatrix-class}
\alias{TargetingMatrix}
\title{S4 class defining a targeting matrix}
\description{
\code{TargetingMatrix} defines a data structure for just the targeting matrix 
(as opposed to the entire \code{TargetingModel})
}
\section{Slots}{

\describe{
\item{\code{.Data}}{matrix.}

\item{\code{numMutS}}{number indicating the number of silent mutations used for 
estimating mutability.}

\item{\code{numMutR}}{number indicating the number of replacement mutations used 
for estimating mutability.}
}}

