% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetingModels.R
\docType{data}
\name{U5N}
\alias{U5N}
\title{Uniform 5-mer null targeting model.}
\format{A \link{TargetingModel} object.}
\usage{
U5N
}
\description{
A null 5-mer model of somatic hypermutation targeting where all substitution, mutability
and targeting rates are uniformly distributed.
}
\seealso{
See \link{HH_S5F} and \link{HKL_S5F} for the human 5-mer targeting models; and 
          \link{MK_RS5NF} for the mouse 5-mer targeting model.
}
\keyword{datasets}
