% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetingModels.R
\docType{data}
\name{U5NModel}
\alias{U5NModel}
\title{Uniform 5-mer targeting model.}
\format{\code{\link{TargetingModel}} object.}
\usage{
U5NModel
}
\description{
A null 5-mer model of somatic hypermutation targeting where all substitution, mutability
and targeting rates are uniformly distributed.
}
\seealso{
See \code{\link{HS5FModel}} the human 5-mer model.
}
\keyword{datasets}

