% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeDB}
\alias{writeDB}
\title{Writes DB to file}
\usage{
writeDB(db, fileName)
}
\arguments{
\item{db}{a data.frame of the DB file.}

\item{fileName}{The filename including path to write to}
}
\description{
Write the contents of DB file to a tab delimited text file. .\cr
}

