% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationProfiling.R
\name{calcObservedMutations}
\alias{calcObservedMutations}
\title{Count the number of observed mutations in a sequence.}
\usage{
calcObservedMutations(inputSeq, germlineSeq, regionDefinition = NULL,
  mutationDefinition = NULL, returnRaw = FALSE, frequency = FALSE)
}
\arguments{
\item{inputSeq}{input sequence. IUPAC ambiguous characters for DNA are supported.}

\item{germlineSeq}{germline sequence. Germline should \strong{not} contain ambiguous
characters.}

\item{regionDefinition}{\link{RegionDefinition} object defining the regions
and boundaries of the Ig sequences. Note, only the part of
sequences defined in \code{regionDefinition} are analyzed.
If NULL, mutations are counted for entire sequence.}

\item{mutationDefinition}{\link{MutationDefinition} object defining replacement
and silent mutation criteria. If \code{NULL} then 
replacement and silent are determined by exact 
amino acid identity.}

\item{returnRaw}{return the positions of point mutations and their corresponding
mutation types, as opposed to counts of mutations across positions.
Also returns the number of bases used as the denominator when 
calculating frequency. Default is \code{FALSE}.}

\item{frequency}{\code{logical} indicating whether or not to calculate
mutation frequencies. The denominator used is the number of bases
that are not one of "N", "-", or "." in either the input or the 
germline sequences. If set, this overwrites \code{returnRaw}. 
Default is \code{FALSE}.}
}
\value{
For \code{returnRaw=FALSE}, an \code{array} with the numbers of replacement (R) 
          and silent (S) mutations. 
          
          For \code{returnRaw=TRUE}, a list containing 
          \itemize{
               \item A data frame (\code{$pos}) whose columns (\code{position}, \code{R}, \code{S}, 
                     and \code{region}) indicate the nucleotide position, the number of R mutations, 
                     the number of S mutations, and the region in which the nucleotide position is 
                     in.
               \item A vector (\code{$nonN}) indicating the number of bases in regions defined by 
                     \code{regionDefinition} (excluding non-triplet overhang, if any) that are not 
                     one of "N", "-", or "." in either the observed or the germline.
          }
          
          For \code{frequency=TRUE}, regardless of \code{returnRaw}, an \code{array} with the 
          frequencies of replacement (R) and silent (S) mutations.
}
\description{
\code{calcObservedMutations} determines all the mutations in a given input seqeunce compared
to its germline sequence.
}
\details{
Each mutation is considered independently in the germline context. Note, only the part of 
\code{inputSeq} defined in \code{regionDefinition} is analyzed. For example, when using 
the default \link{IMGT_V} definition, then mutations in positions beyond 
312 will be ignored. Additionally, non-triplet overhang at the sequence end is ignored.

Only replacement (R) and silent (S) mutations are included in the results. Excluded are: 
\itemize{
     \item Stop mutations
     \item Mutations occurring in codons where one or both of the observed and the 
           germline involve(s) one or more of "N", "-", or ".".
           
           E.g.: the case in which NNN in the germline sequence is observed as NNC in 
           the input sequence.
}
In other words, a result that is \code{NA} or zero indicates absence of R and S mutations, 
not necessarily all types of mutations, such as the excluded ones mentioned above.
}
\examples{
# Use an entry in the example data for input and germline sequence
data(ExampleDb, package="alakazam")
in_seq <- ExampleDb[["SEQUENCE_IMGT"]][100]
germ_seq <-  ExampleDb[["GERMLINE_IMGT_D_MASK"]][100]

# Identify all mutations in the sequence
ex1_raw = calcObservedMutations(in_seq, germ_seq, returnRaw=TRUE)
# Count all mutations in the sequence
ex1_count = calcObservedMutations(in_seq, germ_seq, returnRaw=FALSE)
ex1_freq = calcObservedMutations(in_seq, germ_seq, returnRaw=FALSE, frequency=TRUE)
# Compare this with ex1_count
table(ex1_raw$pos$region, ex1_raw$pos$R)[, "1"]
table(ex1_raw$pos$region, ex1_raw$pos$S)[, "1"]
# Compare this with ex1_freq
table(ex1_raw$pos$region, ex1_raw$pos$R)[, "1"] / ex1_raw$nonN
table(ex1_raw$pos$region, ex1_raw$pos$S)[, "1"] / ex1_raw$nonN

# Identify only mutations the V segment minus CDR3
ex2_raw = calcObservedMutations(in_seq, germ_seq, 
                                regionDefinition=IMGT_V, returnRaw=TRUE)
# Count only mutations the V segment minus CDR3
ex2_count = calcObservedMutations(in_seq, germ_seq, 
                                  regionDefinition=IMGT_V, returnRaw=FALSE)
ex2_freq = calcObservedMutations(in_seq, germ_seq, 
                                 regionDefinition=IMGT_V, returnRaw=FALSE,
                                 frequency=TRUE)
# Compare this with ex2_count
table(ex2_raw$pos$region, ex2_raw$pos$R)[, "1"]
table(ex2_raw$pos$region, ex2_raw$pos$S)[, "1"]                              
# Compare this with ex2_freq
table(ex2_raw$pos$region, ex2_raw$pos$R)[, "1"] / ex2_raw$nonN     
table(ex2_raw$pos$region, ex2_raw$pos$S)[, "1"] / ex2_raw$nonN                                       

# Identify mutations by change in hydropathy class
ex3_raw = calcObservedMutations(in_seq, germ_seq, regionDefinition=IMGT_V,
                                mutationDefinition=HYDROPATHY_MUTATIONS, returnRaw=TRUE)
# Count mutations by change in hydropathy class
ex3_count = calcObservedMutations(in_seq, germ_seq, regionDefinition=IMGT_V,
                                  mutationDefinition=HYDROPATHY_MUTATIONS, returnRaw=FALSE)
ex3_freq = calcObservedMutations(in_seq, germ_seq, regionDefinition=IMGT_V,
                                 mutationDefinition=HYDROPATHY_MUTATIONS, returnRaw=FALSE, 
                                 frequency=TRUE)
# Compre this with ex3_count
table(ex3_raw$pos$region, ex3_raw$pos$R)[, "1"]
table(ex3_raw$pos$region, ex3_raw$pos$S)[, "1"]
# Compare this with ex3_freq
table(ex3_raw$pos$region, ex3_raw$pos$R)[, "1"] / ex3_raw$nonN                                        
table(ex3_raw$pos$region, ex3_raw$pos$S)[, "1"] / ex3_raw$nonN                                        
                                
}
\seealso{
See \link{observedMutations} for counting the number of observed mutations 
in a \code{data.frame}.
}
