% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistToNearest.R
\docType{class}
\name{GmmThreshold-class}
\alias{GmmThreshold-class}
\alias{GmmThreshold}
\alias{print,GmmThreshold-method}
\alias{GmmThreshold-method}
\alias{plot,GmmThreshold,missing-method}
\alias{GmmThreshold-method}
\title{Output of the \code{gmm} method of findThreshold}
\usage{
\S4method{print}{GmmThreshold}(x)

\S4method{plot}{GmmThreshold,missing}(x, y, ...)
}
\arguments{
\item{x}{GmmThreshold object}

\item{y}{ignored.}

\item{...}{arguments to pass to \link{plotGmmThreshold}.}
}
\description{
\code{GmmThreshold} contains output from the \code{gmm} method \link{findThreshold}. 
It includes parameters of two Gaussian fits and threshold cut.
}
\section{Slots}{

\describe{
\item{\code{x}}{input distance vector with NA or infinite values removed.}

\item{\code{omega1}}{first Gaussain mixing proportion.}

\item{\code{omega2}}{second Gaussain mixing proportion.}

\item{\code{mu1}}{first Gaussian mean.}

\item{\code{mu2}}{second Gaussain mean.}

\item{\code{sigma1}}{first Gaussain standard deviation.}

\item{\code{sigma2}}{second Gaussain standard deviation.}

\item{\code{threshold}}{optimum threshold cut.}
}}

\seealso{
\link{findThreshold}
}
