% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistToNearest.R
\docType{class}
\name{GmmThreshold-class}
\alias{GmmThreshold-class}
\alias{GmmThreshold}
\alias{print,GmmThreshold-method}
\alias{GmmThreshold-method}
\alias{plot,GmmThreshold,missing-method}
\alias{GmmThreshold-method}
\title{Output of the \code{gmm} method of findThreshold}
\usage{
\S4method{print}{GmmThreshold}(x)

\S4method{plot}{GmmThreshold,missing}(x, y, ...)
}
\arguments{
\item{x}{GmmThreshold object}

\item{y}{ignored.}

\item{...}{arguments to pass to \link{plotGmmThreshold}.}
}
\description{
\code{GmmThreshold} contains output from the \code{gmm} method \link{findThreshold}. 
It includes parameters of two Gaussian fits and threshold cut.
}
\section{Slots}{

\describe{
\item{\code{x}}{input distance vector with NA or infinite values removed.}

\item{\code{model}}{first-second fit functions.}

\item{\code{cutoff}}{type of threshold cut.}

\item{\code{func1.0}}{mixing weight of the first curve.}

\item{\code{func1.1}}{second parameter of the first curve (either mean of a Normal distribution or shape of a Gamma distribution).}

\item{\code{func1.2}}{third parameter of the first curve (either standard deviation of a Normal distribution or scale of a Gamma distribution).}

\item{\code{func2.0}}{mixing weight of the second curve.}

\item{\code{func2.1}}{second parameter of the second curve (either mean of a Normal distribution or shape of a Gamma distribution)..}

\item{\code{func2.2}}{third parameter of the second curve (either standard deviation of a Normal distribution or scale of a Gamma distribution).}

\item{\code{loglk}}{the fit log-likelihood.}

\item{\code{threshold}}{the threshold cut.}

\item{\code{sensitivity}}{the sensitivity.}

\item{\code{specificity}}{the specificity.}

\item{\code{pvalue}}{the p-value from Hartigans' dip statistic (HDS) test, with values less than 0.05 indicating significant bimodality.}
}}

\seealso{
\link{findThreshold}
}
