% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetingModels.R
\docType{data}
\name{HS5FModel}
\alias{HS5FModel}
\title{Human 5-mer targeting model.}
\format{\code{\link{TargetingModel}} object.}
\usage{
HS5FModel
}
\description{
5-mer model of somatic hypermutation targeting based on analysis of silent mutations
in functional Ig sequences from Homo sapiens.
}
\references{
\enumerate{
  \item  Yaari G, et al. Models of somatic hypermutation targeting and substitution based 
           on synonymous mutations from high-throughput immunoglobulin sequencing data. 
           Front Immunol. 2013 4(November):358.
 }
}
\seealso{
See \code{\link{HS1FDistance}} for the 1-mer distance matrix from the same 
          publication.
}
\keyword{datasets}

