% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionDefinitions.R
\name{IMGT_SCHEMES}
\alias{IMGT_SCHEMES}
\alias{IMGT_V}
\alias{IMGT_V_BY_REGIONS}
\title{IMGT unique numbering schemes}
\format{A \link{RegionDefinition} object containing:
\itemize{
  \item  \code{IMGT_V}:                     Grouped CDR and FWR V-segment regions excluding CDR3.
  \item  \code{IMGT_V_BY_REGIONS}:          Individual CDR and FWR V-segment regions excluding CDR3.
}}
\description{
Definitions of the CDR and FWR, according to the IMGT unique numbering scheme.
}
\references{
\enumerate{
  \item  Lefranc MP, et al. IMGT unique numbering for immunoglobulin and T cell 
           receptor variable domains and Ig superfamily V-like domains. 
           Developmental and comparative immunology. 2003 27:55-77.
}
}
