% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationDefinitions.R
\name{MUTATION_SCHEMES}
\alias{MUTATION_SCHEMES}
\alias{CHARGE_MUTATIONS}
\alias{HYDROPATHY_MUTATIONS}
\alias{POLARITY_MUTATIONS}
\alias{VOLUME_MUTATIONS}
\title{Amino acid mutation definitions}
\format{A \link{MutationDefinition} object defining:
\itemize{
  \item  \code{CHARGE_MUTATIONS}:      Amino acid mutations are defined by changes
                                       in side chain charge class.
  \item  \code{HYDROPATHY_MUTATIONS}:  Amino acid mutations are defined by changes
                                       in side chain hydrophobicitity class.
  \item  \code{POLARITY_MUTATIONS}:    Amino acid mutations are defined by changes
                                       in side chain polarity class.
  \item  \code{VOLUME_MUTATIONS}:      Amino acid mutations are defined by changes
                                       in side chain volume class.
}}
\description{
Definitions of replacement (R) and silent (S) mutations for different amino acid
physicochemical classes.
}
\references{
\enumerate{
  \item  \url{http://www.imgt.org/IMGTeducation/Aide-memoire/_UK/aminoacids/IMGTclasses.html} 
}
}
