% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotSelection}
\alias{plotSelection}
\title{Plot the results of bootstrapDiversity}
\usage{
plotSelection(db, groups = NULL, colors = NULL,
  main_title = "Selection Strength", legend_title = NULL)
}
\arguments{
\item{db}{data.frame returned by calcGroupedBaseline}

\item{groups}{column names by which you want the anlaysis grouped by}

\item{colors}{values are colors to assign to those group values}

\item{main_title}{string specifying the plot title.}

\item{legend_title}{string specifying the legend title.}
}
\value{
a \code{ggplot} object
}
\description{
\code{plotSelection} plots.
}

